chore: clean up display structs
