--- src/display/traits.rs
@@ -1,9 +1,0 @@
-/// All commands need to have this trait which gives the address of the command
-/// which needs to be send via SPI with activated CommandsPin (Data/Command Pin in CommandMode)
-pub(crate) trait Command {
-    fn address(self) -> u8;
-}
-
-pub trait IsBusy {
-    fn is_busy(&mut self) -> bool;
-}
