+++ src/lib.rs
@@ -7,3 +7,9 @@ pub mod display;
 pub mod shift_register;
 
 pub use display::InkyFrameDisplay;
+
+/// Trait for determining if the e-ink display is busy
+/// This could be the busy_pin for the inky impression or the shift_register for inky frame
+pub trait IsBusy {
+    fn is_busy(&mut self) -> bool;
+}
